refactor: extract model validation mixin, admin handlers, split math_tools.py
- Extracted _model_validation_mixin.py from legacy_server.py (157 lines)
- Extracted _admin_handlers_mixin.py from legacy_server.py (78 lines)
- Split math_tools.py (932 lines) into math_tools/ package:
  - _executor.py (402 lines) - SafePythonExecutor + helpers
  - _solver.py (134 lines) - SelfVerificationSolver
  - __init__.py - re-exports for backward compat
